Web applications run through web browsers like Google Chrome. The program runs on a web server, rather than on the PC, or local server for traditional applications. Web application pages interact and respond with users requests, unlike basic website pages where pages are all pre-formatted. Web applications can utilize a database to store permanent information such as product descriptions/costs, and customer orders.
The website is developed on real life. It is very helpful in business applications. Today’s extremely exhausting work environment dictates that individuals require some joyful holiday. The website will provide a stress-free joyful refreshing holidays with cost competitive and customized packages according to their requirements. As it is web based, it can provide services in almost each and every city in India. A traveler can find everything related to travelling services under one roof by this website. The site will help to estimate the benefits and calculate the requirements very easily. The benefits will be in terms of speedy query, smooth and faster operations at all levels as it implements the UI/UX design. Maintaining the database as required. Database can be modified accordingly.